Move LayerClustersAlgo to be Plugins
